+++ src/App/Modules/AppModule.js
@@ -6,14 +6,14 @@ import WebView from './WebView';
  *
  * @param {*} { module }
  */
-const PageModule = ({ module }) => (
+const AppModule = ({ module }) => (
   <WebView
     /* Set key attribute here so that the WebView instance will be reset and 
       componentDidMount will be called when it navigates to another module of the same type */
-    key={module.name}
+    key={module.info.name}
     module={module}
     style={{ width: '100%', height: '100%' }}
   />
 );
 
-export default PageModule;
+export default AppModule;