+++ src/App/Navigation/index.js
@@ -1,33 +1,24 @@
-// External Dependencies
 import React from 'react';
 import { connect } from 'react-redux';
 import styled from '@emotion/styled';
 import { keyframes } from '@emotion/core';
 
-// Internal Global Depnedencies
 import HorizontalLine from 'components/HorizontalLine';
 import Icon from 'components/Icon';
 import Tooltip from 'components/Tooltip';
 import ModuleIcon from 'components/ModuleIcon';
-import { getActiveModules } from 'lib/modules';
 import { consts, timing } from 'styles';
+import { legacyMode } from 'consts/misc';
 
-// Internal Local Dependencies
-import NavLinkItem from './NavLinkItem';
-
-// Images
 import logoIcon from 'icons/logo.svg';
 import sendIcon from 'icons/send.svg';
-import chartIcon from 'icons/chart.svg';
 import transactionsIcon from 'icons/transaction.svg';
 import addressBookIcon from 'icons/address-book.svg';
 import settingsIcon from 'icons/settings.svg';
 import consoleIcon from 'icons/console.svg';
 import userIcon from 'icons/user.svg';
-import tokenIcons from 'icons/blockexplorer-invert-white.svg';
-import { legacyMode } from 'consts/misc';
-// import shapeshiftIcon from 'icons/shapeshift.svg';
-// import trustListIcon from 'icons/trust-list.svg';
+
+import NavLinkItem from './NavLinkItem';
 
 __ = __context('NavigationBar');
 
@@ -79,24 +70,19 @@ const NavItem = ({ icon, children, ...rest }) => (
  * @memberof Navigation
  */
 const ModuleNavItem = ({ module }) => (
-  <Tooltip.Trigger tooltip={module.displayName} position="top">
-    <NavLinkItem to={`/Modules/${module.name}`}>
+  <Tooltip.Trigger tooltip={module.info.displayName} position="top">
+    <NavLinkItem to={`/Modules/${module.info.name}`}>
       <ModuleIcon module={module} />
     </NavLinkItem>
   </Tooltip.Trigger>
 );
 
-const ModuleNavItems = connect(
-  state => ({
-    modules: getActiveModules(state.modules, state.settings.disabledModules),
-  }),
-  null,
-  null
-  // { pure: false }
-)(({ modules }) =>
-  modules
-    .filter(module => module.type === 'app')
-    .map(module => <ModuleNavItem key={module.name} module={module} />)
+const ModuleNavItems = connect(state => ({
+  modules: state.modules,
+}))(({ modules }) =>
+  Object.values(modules)
+    .filter(module => module.enabled && module.info.type === 'app')
+    .map(module => <ModuleNavItem key={module.info.name} module={module} />)
 );
 
 /**
@@ -126,10 +112,6 @@ const Navigation = () => (
 
       <NavItem icon={transactionsIcon} to="/Transactions">
         {__('Transactions')}
-      </NavItem>
-
-      <NavItem icon={chartIcon} to="/Market">
-        {__('Market Data')}
       </NavItem>
 
       <NavItem icon={addressBookIcon} to="/AddressBook">