+++ src/index.js
@@ -3,17 +3,30 @@ import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 
 import 'appMenu';
+import 'lib/market';
 import store from 'store';
 import { walletEvents } from 'lib/wallet';
+import { startCore } from 'lib/core';
+import initialSettings from 'data/initialSettings';
 import App from './App';
 
-walletEvents.emit('pre-render');
+async function run() {
+  try {
+    if (!initialSettings.manualDaemon) {
+      await startCore();
+    }
+  } finally {
+    walletEvents.emit('pre-render');
 
-render(
-  <Provider store={store}>
-    <App />
-  </Provider>,
-  document.getElementById('root')
-);
+    render(
+      <Provider store={store}>
+        <App />
+      </Provider>,
+      document.getElementById('root')
+    );
 
-walletEvents.emit('post-render');
+    walletEvents.emit('post-render');
+  }
+}
+
+run();